feat: new code organization
